add insert dylib tools

+++ Tools/mpack.sh
@@ -19,6 +19,9 @@ echo "packing..."
 monkeyparser="$MONKEYDEV_PATH/bin/monkeyparser"
 substrate="$MONKEYDEV_PATH/MFrameworks/libsubstitute.dylib"
 
+# insert_dylib
+INSERT_DYLIB="${MONKEYDEV_PATH}/bin/insert_dylib"
+
 #exename
 TARGET_APP_PATH=$(find "$SRCROOT/$TARGET_NAME/TargetApp" -type d | grep ".app$" | head -n 1)
 
@@ -46,9 +49,13 @@ fi
 #insert dylib
 BUILD_DYLIB_PATH="$BUILT_PRODUCTS_DIR/lib$TARGET_NAME.dylib"
 
-if [[ ! -f "$APP_BINARY_PATH".insert ]]; then
+if [[ ! -f "$APP_BINARY_PATH".insert ]] && [[ ${MONKEYDEV_INSERT_DYLIB} == "YES" ]]; then
 	cp -rf "$substrate" "$TARGET_APP_PATH/Contents/MacOS/"
-	"$monkeyparser" install -c load -p "@executable_path/lib$TARGET_NAME.dylib" -t "$APP_BINARY_PATH"
+	if [[ ${MONKEYDEV_INSERT_DYLIB_TOOLS} == "INSERT_DYLIB" ]];then
+		"$INSERT_DYLIB" --inplace --overwrite --all-yes "@executable_path/lib$TARGET_NAME.dylib" "$APP_BINARY_PATH"
+	else
+		"$monkeyparser" install -c load -p "@executable_path/lib$TARGET_NAME.dylib" -t "$APP_BINARY_PATH"
+	fi
 	echo "insert" >> "$APP_BINARY_PATH".insert
 fi
